Revert "Знакомство с HTTP API клиентами. Практика: Написание HTTP API клиента."
