--- HW/7.3_HW.md
@@ -1,61 +1,0 @@
-**7.3 Знакомство с HTTP API клиентами**
-
-Практика: Написание HTTP API клиента
-
-В данном задании вам самостоятельно необходим реализовать API клиент для работы с эндпоинтами /api/v1/cards сервиса http-gateway 
-
-Требования к реализации
-Создайте класс CardsGatewayHTTPClient, который:
-
-Должен быть унаследован от HTTPClient.
-Должен содержать методы: issue_virtual_card_api, issue_physical_card_api.
-Должен быть задокументирован с помощью docstring.
-Метод issue_virtual_card_api должен:
-
-Выполнять POST-запрос к эндпоинту /api/v1/cards/issue-virtual-card для создания виртуальной карты.
-Быть полностью аннотирован:
-Входной аргумент request должен использовать TypedDict с корректной структурой данных.
-Возвращаемое значение должно быть аннотировано как httpx.Response.
-Содержать docstring с описанием назначения метода, входных параметров и возвращаемого значения.
-Метод issue_physical_card_api должен:
-Выполнять POST-запрос к эндпоинту /api/v1/cards/issue-physical-card для создания физической карты.
-Быть полностью аннотирован:
-Входной аргумент request должен использовать TypedDict с корректной структурой данных.
-Возвращаемое значение должно быть аннотировано как httpx.Response.
-Содержать docstring с описанием назначения метода, входных параметров и возвращаемого значения.
-Файл для реализации:
-
-API клиент должен быть размещен в файле:
-/clients/http/gateway/cards/client.py
-
-                  
-Документирование кода:
-
-Все классы и методы должны быть задокументированы с использованием docstring.
-
-В docstring методов необходимо описать:
-
-Назначение метода.
-Описание входных параметров.
-Формат возвращаемого значения.
-Структура расположения файла client.py:
-```
-└── performance-tests/
-    └── clients/
-        └── http/
-            ├── gateway/
-            │   ├── __init__.py
-            │   └── cards/
-            │       ├── __init__.py
-            │       └── client.py
-            ├── __init__.py
-            └── client.py
-
-```
-
-                  
-Требования к коду:
-Читаемость и структурированность.
-Соблюдение стандартов аннотации типов.
-Корректное именование классов и методов.
-Обязательное документирование кода с помощью docstring.
