+++ httpx_get_documents.py
@@ -1,7 +1,7 @@
 '''
 Сценарий, который мы реализуем
 В рамках текущего задания вы реализуете скрипт, выполняющий следующую последовательность:
-    Создание пользователя. POST /api/v1/cards
+    Создание пользователя. POST /api/v1/users
     Открытие счёта с кредитной картой. POST /api/v1/accounts/open-credit-card-account. В процессе открытия счёта
         автоматически создаются и загружаются связанные с ним документы (тариф и контракт) в MinIO.
     Получение тарифного документа по счёту. GET /api/v1/documents/tariff-document/{account_id}
